[passwords] feat: Implement hashing
