+++ backend/package.json
@@ -40,6 +40,8 @@
   },
   "dependencies": {
     "@prisma/client": "5.7.1",
+    "@types/bcryptjs": "^2.4.6",
+    "bcryptjs": "^2.4.3",
     "cors": "2.8.5",
     "dotenv": "16.3.2",
     "express": "4.18.2",
